backend/app/requirements.txt
Flask==1.0.2
Flask-SocketIO==3.0.1
eventlet==0.24.1
websocket-client==0.52.0
